backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.16.5
altair==5.5.0
aniso8601==10.0.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.11.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bitarray==3.4.1
bleach==6.2.0
blinker==1.7.0
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botframework-connector>=4.14.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.1.1
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7.1
contourpy==1.3.2
cryptography==45.0.6
curl_cffi==0.11.4
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
EasyProcess==1.1
elevenlabs==0.2.27
entrypoint2==1.1
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.13
fastjsonschema==2.21.1
fdb==2.0.3
filelock==3.18.0
filetype==1.2.0
firebird-base==2.0.0
firebird-driver==2.0.2
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2025.5.1
gevent==25.5.1
geventhttpclient==2.3.4
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
griffe==1.7.3
gunicorn==21.2.0
h11==0.14.0
hexbytes==1.3.1
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.1
huggingface-hub==0.33.0
idna==3.6
imbalanced-learn==0.13.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-experimental==0.3.4
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.27
llama-cloud==0.1.28
llama-cloud-services==0.6.36
llama-index==0.12.43
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.43
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-instrumentation==0.2.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==0.2.2
llama-parse==0.6.36
locust==2.37.10
locust-cloud==1.23.2
lxml==5.4.0
Mako==1.3.10
MarkupSafe==2.1.4
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.6
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.1
mss==10.1.0
multidict==6.0.4
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.5
openai==2.1.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.3
pdf2image==1.17.0
peewee==3.18.1
pexpect==4.9.0
Pillow==9.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pit==0.5
platformdirs==4.3.8
playwright==1.54.0
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py-mon==2.1.0
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.17.2
PyJWT==2.10.1
pymon==1.1
pymongo==4.13.2
pynput==1.8.1
pyparsing==3.2.3
pypdf==5.6.1
pyscreenshot==3.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.13.0
python-telegram-bot==20.0a4
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rlp==4.1.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.34.2
selenium-stealth==1.0.6
sentence-transformers==4.1.0
simple-websocket==1.1.0
simplejson==3.20.1
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
tqdm==4.66.1
traitlets==5.14.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
tushare==1.4.21
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
web3==7.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.2
xxhash==3.6.0
yarl==1.9.4
yfinance==0.2.63
zope.event==5.0
zope.interface==7.2
zstandard==0.24.0
